usage-tracker/requirements.txt
redis==4.5.4
psycopg2-binary==2.9.6
pydantic-settings
python-dotenv